ghgfh

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: mrk <marvin@42.fr>                         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 17:32:26 by mrk               #+#    #+#             */
-/*   Updated: 2024/03/11 17:29:12 by mrk              ###   ########.fr       */
+/*   Updated: 2024/03/11 17:29:54 by mrk              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -23,7 +23,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t size)
 		return (0);
 	while (dst[dst_len] && dst_len < (int)size)
 		dst_len++;
-	if (dst_len < size)
+	if (dst_len < (int)size)
 		len = dst_len + ft_strlen((char *)src);
 	else
 		return (size + ft_strlen((char *)src));
